Split filters into multiple files
