+++ src/include/SIMD.h
@@ -1,0 +1,3 @@
+#pragma once
+
+enum class SIMD { Scalar, AVX512 };
